Delete asset,, Texture now. new default window size, scalable with float.

+++ Soli-Rom Engine/src/Engine/App.h
@@ -6,7 +6,7 @@ namespace SoliRom
 	class App
 	{
 	public:
-		App(int _windowSizeMul);
+		App(float _windowSizeMul);
 		~App();
 
 		void Run();
